requirements.txt
ttkthemes~=3.2.2
Mako~=1.3.6
PyYAML~=6.0.2
lxml~=5.3.0
markdownify~=0.14.1
setuptools~=75.8.0
PyMuPDF~=1.25.3
pypdf~=5.2.0